requirements.txt
# requirements.txt
feedparser
newsapi-python
ollama
pandas
requests
streamlit
pycoingecko
plotly
python-dotenv
beautifulsoup4